update documentation and refactoring
