+++ packages/kafka/src/processMessage/lyric/handleDocument.ts
