+++ packages/kafka/src/processMessage/song/handleDocument.ts
